+++ src/lib.rs
@@ -17,8 +17,8 @@
 //! use uom::si::time::second;
 //!
 //! fn main() {
-//! 	let length = Length::new(5.0, kilometer);
-//! 	let time = Time::new(15.0, second);
+//! 	let length = Length::new::<kilometer>(5.0);
+//! 	let time = Time::new::<second>(15.0);
 //! 	let _velocity = length / time;
 //! 	//let error = length + time; // error[E0308]: mismatched types
 //! }